+++ src/kudu/common/schema.h
@@ -558,9 +558,8 @@ class Schema {
   // result will have fewer columns than requested.
   //
   // The resulting schema will have no key columns defined.
-  Status CreateProjectionByIdsIgnoreMissing(
-      const std::vector<int>& col_ids,
-      Schema* out) const;
+  Status CreateProjectionByIdsIgnoreMissing(const std::vector<size_t>& col_ids,
+                                            Schema* out) const;
 
   // Encode the key portion of the given row into a buffer
   // such that the buffer's lexicographic comparison represents
