+++ src/kudu/tserver/tablet_server-test.cc
@@ -854,8 +854,8 @@ TEST_F(TabletServerTest, TestKUDU_176_RecoveryAfterMajorDeltaCompaction) {
   {
     vector<shared_ptr<tablet::RowSet> > rsets;
     tablet_peer_->tablet()->GetRowSetsForTests(&rsets);
-    vector<int> col_ids = { tablet_peer_->tablet()->schema()->column_id(1),
-                            tablet_peer_->tablet()->schema()->column_id(2) };
+    vector<size_t> col_ids = { tablet_peer_->tablet()->schema()->column_id(1),
+                               tablet_peer_->tablet()->schema()->column_id(2) };
     ASSERT_OK(tablet_peer_->tablet()->DoMajorDeltaCompaction(col_ids, rsets[0]))
   }
 
@@ -888,8 +888,8 @@ TEST_F(TabletServerTest, TestKUDU_177_RecoveryOfDMSEditsAfterMajorDeltaCompactio
   {
     vector<shared_ptr<tablet::RowSet> > rsets;
     tablet_peer_->tablet()->GetRowSetsForTests(&rsets);
-    vector<int> col_ids = { tablet_peer_->tablet()->schema()->column_id(1),
-                            tablet_peer_->tablet()->schema()->column_id(2) };
+    vector<size_t> col_ids = { tablet_peer_->tablet()->schema()->column_id(1),
+                               tablet_peer_->tablet()->schema()->column_id(2) };
     ASSERT_OK(tablet_peer_->tablet()->DoMajorDeltaCompaction(col_ids, rsets[0]));
   }
   // Verify that data is still the same.
